Helper -> Core refactoring
